fix: fix styling and wording to match the design refactor: move and rename component
