+++ packages/shared/components/modals/NetworkSummaryModal.svelte
