+++ packages/shared/components/modals/index.js
@@ -1,5 +1,5 @@
 export { default as AccountActionsModal } from './AccountActions.svelte'
 export { default as AccountSwitcherModal } from './AccountSwitcher.svelte'
 export { default as Modal } from './Modal.svelte'
-export { default as NetworkIndicatorModal } from './NetworkIndicator.svelte'
+export { default as NetworkSummaryModal } from './NetworkSummaryModal.svelte'
 export { default as ProfileActionsModal } from './ProfileActions.svelte'
