+++ packages/shared/locales/en.json
@@ -1315,7 +1315,7 @@
             "warningText": "You are currently using a developer profile and are connected to {networkName}"
         },
         "networkIndicator": {
-            "warningText": "Your connected network {networkName} is having issues"
+            "warningText": "Network performance degraded. Message confirmation may take a bit longer than usual."
         }
     },
     "permissions": {
